requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv_python==4.10.0.84
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
utils==1.0.2
scikit-learn
openpyxl
starlette
reportlab
uvicorn[standard]
isort



# Using yolov11 model
ultralytics==8.3.27

# Tensorflow
tensorflow==2.18.0
keras==3.6.0
joblib==1.4.2

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop~=0.1.1.post2209072238
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

telepot~=12.7
torchmetrics~=0.10.2


# CUDA
<<<<<<< HEAD
nvidia-cudnn-cu12>=9.5.1
=======
nvidia-cudnn-cu12>=9.1.0
>>>>>>> 6708c07 (integrate w software)
